requirements.txt
anyio==3.5.0
asgiref==3.5.0
black==22.3.0
click==8.1.2
colorama==0.4.4
fastapi==0.75.1
flake8==4.0.1
h11==0.13.0
idna==3.3
importlib-metadata==4.2.0
mccabe==0.6.1
mypy-extensions==0.4.3
mysk-utils @ git+https://github.com/suankularb-wittayalai-school/mysk-utils.git@main
pathspec==0.9.0
platformdirs==2.5.1
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
python-dotenv==0.20.0
sniffio==1.2.0
starlette==0.17.1
tomli==2.0.1
typed-ast==1.5.2
typing-extensions==4.1.1
uvicorn==0.17.6
zipp==3.8.0